requirements.txt
django==4.1.3
python-dotenv==0.20.0
psycopg2==2.9.5
djangorestframework==3.14.0
drf_spectacular==0.24.2
djoser==2.1.0
celery==5.2.7
django-celery-results==2.4.0
redis==4.3.4
django-celery-beat==2.4.0
pillow==9.3.0
qrcode==7.3.1
django-nested-admin==4.0.0
django-nested-inline==0.4.5
pydantic==1.10.2



